faster faster and faster

+++ src/std.c
@@ -15,8 +15,8 @@ __asm__ volatile ("rep movsb"
 : "memory");
 }
 
-void kmemset(uint8_t *dest, uint8_t val, uint32_t len) {
-    uint8_t *temp = (uint8_t *)dest;
+void kmemset(uint32_t *dest, uint32_t val, uint32_t len) {
+    uint32_t *temp = (uint32_t *)dest;
     for ( ; len != 0; len--) *temp++ = val;
 }
